+++ packages/terra-profile-image/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Fixed
+  * Fixed broken link in documentation.
+
 ## 3.41.1 - (January 4, 2022)
 
 * Changed