+++ packages/terra-profile-image/CHANGELOG.md
@@ -4,6 +4,9 @@
 
 ## 3.41.1 - (January 4, 2022)
 
+* Changed 
+  * When `src` is not provided Aria-hidded is true.
+
 * Changed
   * Migrate docs to terra-core-docs.