+++ packages/terra-profile-image/CHANGELOG.md
@@ -5,7 +5,7 @@
 ## 3.41.1 - (January 4, 2022)
 
 * Changed 
-  * When `src` is not provided Aria-hidded is true.
+  * Minor dependency version bump
 
 * Changed
   * Migrate docs to terra-core-docs.