+++ packages/terra-profile-image/CHANGELOG.md
@@ -2,10 +2,10 @@
 
 ## Unreleased
 
-## 3.41.1 - (January 4, 2022)
-
 * Changed 
-  * Minor dependency version bump
+  * Added `role = presenetation` for Decorative profile-image.
+
+## 3.41.1 - (January 4, 2022)
 
 * Changed
   * Migrate docs to terra-core-docs.